Добавлен слой дто, вынесена обработка ошибок, добавлен фильтр

+++ src/main/java/model/ExchangeRate.java
@@ -1,14 +1,16 @@
 package model;
 
+import java.math.BigDecimal;
+
 public class ExchangeRate {
     private final int _id;
     private final Currency _baseCurrency;
     private final Currency _targetCurrency;
-    private final double _rate;
+    private final BigDecimal _rate;
 
 
 
-    public ExchangeRate(int id, Currency baseCurrency, Currency targetCurrency, double rate){
+    public ExchangeRate(int id, Currency baseCurrency, Currency targetCurrency, BigDecimal rate){
         _id = id;
         _baseCurrency = baseCurrency;
         _targetCurrency = targetCurrency;
@@ -29,6 +31,6 @@ public class ExchangeRate {
     }
 
 
-    public double getRate() { return _rate; }
+    public BigDecimal getRate() { return _rate; }
 
 }
